Fixed empty zone name bug
For some reason, EVENT_ZONE_CHANGED passes in an empty zoneName string. Either a bug, or it's not used the way I thought it did.
TEB.getZone now does what getZoneRebuild did (i.e. what TEB.zone originally did) to get the zone name. Not as elegant, but it works.